+++ NEWS.md
@@ -1,3 +1,4 @@
 # xpose.nlmixr 0.2.0
 
 * Added a `NEWS.md` file to track changes to the package.
+* Updated functions to reflect changes in `tibble` and `dplyr`.